move services and their tests to new sub packages
