+++ src/main/java/com/example/myTestApp/service/counter/NumOfDoublesCounter.java
@@ -1,4 +1,4 @@
-package com.example.myTestApp.service;
+package com.example.myTestApp.service.counter;
 
 public interface NumOfDoublesCounter {
     int count (String str);
