+++ src/main/java/com/example/myTestApp/service/evaluator/ArithmeticExpressionEvaluator.java
@@ -1,4 +1,4 @@
-package com.example.myTestApp.service;
+package com.example.myTestApp.service.evaluator;
 
 import org.springframework.stereotype.Service;
